+++ src/components/Inputs/Input.jsx
@@ -1,0 +1,6 @@
+
+export default function Input({autocomplete,id,name,type,className,placeholder}){
+    <div>
+        <input autocomplete={autocomplete} id = {id} name={name} type={type} className={className} placeholder={placeholder} />
+    </div>
+}
