+++ src/components/Login-Sign/LoginSign.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+import Layout from "./LoginLayout";
+import { Child } from "./LoginChild";
+
+export default function Login() {
+  return <Layout child={<Child />} />;
+}
