+++ src/components/LoginSignButtons.jsx/LoginSignButtons.jsx
@@ -1,0 +1,9 @@
+export default function Button({ onClick, value, name }) {
+  return (
+    <div>
+      <button onClick={onClick} value={value} name={name}>
+        {name}
+      </button>
+    </div>
+  );
+}
